requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
conda==4.8.3
conda-package-handling==1.7.0
cryptography==2.9.2
Deprecated==1.2.10
ERPpeek==1.7.1
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycosat==0.6.3
pycparser==2.20
PyGithub==1.53
PyJWT==1.7.1
pyOpenSSL==19.1.0
PySocks==1.7.1
requests==2.23.0
ruamel-yaml==0.15.87
six==1.14.0
tqdm==4.46.0
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.12.1
